+++ apps/api/src/projects/guards/project-access-own.guard.ts
@@ -1,12 +1,14 @@
 import { CanActivate, ExecutionContext, Injectable } from "@nestjs/common";
 import { Reflector } from "@nestjs/core";
+import { Request } from "express";
+import { ProjectId } from "src/components/guards/apikey.guard";
 import { CorrelationLoggerService } from "../../common/services/logger.service";
 import { ProjectsService } from "../projects.service";
 
 export const ProjectIdParameterKey = Reflector.createDecorator<string>({});
 
 /** Makes sure that the project being accessed belongs to the API key making the
- * request. Stores the current project ID in `request.projectId`
+ * request. Stores the current project ID in `request[ProjectId]`
  *
  * If the parameter name is not `'id'`, then use the ProjectIdParameterKey
  * decorator to specify the parameter name.
@@ -20,7 +22,7 @@ export class ProjectAccessOwnGuard implements CanActivate {
   ) {}
 
   async canActivate(context: ExecutionContext): Promise<boolean> {
-    const request = context.switchToHttp().getRequest();
+    const request: Request = context.switchToHttp().getRequest();
     const correlationId = request["correlationId"];
     const apiKey = request.headers["x-api-key"];
 
@@ -36,19 +38,30 @@ export class ProjectAccessOwnGuard implements CanActivate {
         ProjectIdParameterKey,
         context.getHandler(),
       );
+      const authorizedProjectId = request[ProjectId];
+      if (!authorizedProjectId) {
+        this.logger.warn(
+          `[${correlationId}] No project ID provided for API key ${apiKey}`,
+        );
+        return true; // Allow the request to proceed, let the controller handle missing projectId
+      }
 
       const projectId = projectIdParameterKey
         ? request.params[projectIdParameterKey]
         : request.params.id;
-
-      // Store the project ID in the request for use in controllers
-      request.projectId = projectId;
 
       if (!projectId) {
         this.logger.warn(
           `[${correlationId}] No project ID provided for API key ${apiKey}`,
         );
         return true; // Allow the request to proceed, let the controller handle missing projectId
+      }
+
+      if (projectId !== authorizedProjectId) {
+        this.logger.warn(
+          `[${correlationId}] Project ID ${projectId} does not match authorized project ID ${authorizedProjectId}`,
+        );
+        return false;
       }
 
       const project = await this.projectsService.findOne(projectId);